Improved download dialogue.
Added the ability to specify a form of compression, either as GZIP or
ZIP, which the resulting file is wrapped in. Also streamline the
slightly confusing footer on the download dialogue to highlight the main
action (downloading) and clearly indicate that Galaxy is an alternative.
Additionally added support for saving the user's galaxy preference where
supported in the intermine API.