Fixed subtle debinding bug.
The events bound to the popover content were unbound on node removal
(due to jQuery's implementation of maintaining a mapping between element
and callback list). The solution is to detach on hide, to prevent this
clearing happening on the next dom insert, and this requires listening
for the hide event.