Changes to do with icons.
This changeset reverts back to bootstrap 2.3.2 icons glyphicon icons
(which are png files set as back-ground images). This is due to
compatibility issues with ie-11 that were discovered relating to the use
of web-fonts as employed by modern glyphicons and font-awesome.

This changeset allows for the end-user to choose which icon set they
want (defaulting to 'glyphicons') by setting the value of
`intermine.options.Style.icons`. If set to `fontawesome`, then
fontawesome icons will be loaded.