Raised checkbox helper to full sub-component.
It behaves like an angular checkbox - watching and setting its
own model - this removes the need for duplicate boilerplate DOM events,
and simplifies data-flow. It also enforces some stylistic consistency.

It can need some styling in certain contexts (such as the facet-rows),
because it receives 10px of margin above and below by default (not
always desirable).