Fixed initial expansion.
This had nothing to do with fragments, but everything to do with the
render cycle - the main view was set to re-render on model changes, but
this meant that the subtable child was removed and not replaced due to
the 'rendered' check in renderTable. This is an example of the truism
that cache invalidation is one of the hard problems.