Skipping replaced paths works
And is considerably more elegant than it ever was before! Rather than
building up two different indices (processed and replaced by) we build
a single one (skipped) that lets us know which cells we can skip. We
only loop once and we cache the skip list generation so we only
generate it once.