Added script to configure css.
This should suffice for the simple cases where all that is required is
to adjust the variables. It should also, tbh, work well for more complex
cases, as it does not preclude the extension from loading its own
includes or complex subsystems. In fact you could even use this as a
mechanism for merging two less trees. While feasible, that would not be
a good use of this, and it would be better to include the imtables css
in a consuming project (as imtables does for bootstrap and fontawesome).